==> real_estate/v4/00_fed_df.ipynb <==
import pandas as pd

dfs = {
    "01": pd.read_csv("~/real_estate/data/raw/fred_3monthtreasurybill.csv"),
    "02": pd.read_csv("~/real_estate/data/raw/fred_average_hourly_earnings_seasonally_adjusted.csv"),
    "03": pd.read_csv("~/real_estate/data/raw/fred_fixed_mortgage_rate.csv"),
    "04": pd.read_csv("~/real_estate/data/raw/fred_total_nonfarm_payroll.csv")
}

variables = {
    "01": "90_day_treasury_bill_rate",
    "02": "mean_hourly_earnings_adj",
    "03": "mean_fixed_mortgage_rate",
    "04": "total_nonfarm_payroll"
}

dataframes_melt = {}

for i in variables.keys():
    
    varname = dfs[i].columns[1]
    df_ = dfs[i].rename(columns={"DATE":"date",varname:variables[i]})
    df_["date"] = [j[:7] for j in df_["date"].values]
    df_ = df_.groupby(["date"], as_index=False).agg('mean')
    dataframes_melt[i] = df_

# Merge with Fed dfs
df = pd.merge(dataframes_melt["01"], dataframes_melt["02"], on=["date"], how="left")
for i in ["03","04"]:
    df = pd.merge(df, dataframes_melt[i], on=["date"], how="left")

# Filter dates, remove NaNs
df = df.loc[df["date"] > '2000']
df = df.dropna()

# Date format
df["month"] = [int(i[5:]) for i in df["date"].values]
df["year"] = [int(i[:4]) for i in df["date"].values]
df = df.drop(["date"], axis=1)

df.head()

df.to_csv("~/real_estate/data/v4/fed_dataframe.csv",index=False)
