==> real_estate/v4/01_merge_dfs.ipynb <==
import pandas as pd

# Load data
yelp_df = pd.read_csv("~/real_estate/data/v4/yelp_zipcode_data.csv")
zillow_df = pd.read_csv("~/real_estate/data/v4/zillow_zipcode_data.csv")
irs_df = pd.read_csv("~/real_estate/data/v4/irs_zipcode_data.csv")
fed_df = pd.read_csv("~/real_estate/data/v4/fed_dataframe.csv")
state_df = pd.read_csv("~/real_estate/data/raw/zip_code_state.csv")

# Filter out Canadian zip codes
yelp_df = yelp_df.loc[yelp_df["postal_code"].isin(set(irs_zip))]

# Merge
df = pd.merge(zillow_df, yelp_df, on=["postal_code","year","month"], how="outer")
df = df.dropna(subset=["zhvi_condo"])
df = pd.merge(df, irs_df, on=["postal_code","year"], how="left")
df = pd.merge(df, fed_df, on=["year","month"], how="left")

# Format
df = df.loc[df["year"] >= 2010]
df = df.dropna(subset=["state"])
df = df.sort_values(by=["postal_code","year","month"])

# Get yelp stuff
yelp_df = df.dropna(subset=["mean_rating"])

# Remove columns with tons of nas
df = df.drop(["zhvi_1bed","zhvi_2bed","zhvi_3bed"], axis=1)
yelp_df = yelp_df.drop(["zhvi_1bed","zhvi_2bed","zhvi_3bed"], axis=1)

# Export
df.to_csv("~/real_estate/data/v4/unsupervised_df.csv",index=False)
yelp_df.to_csv("~/real_estate/data/v4/yelp_unsupervised_df.csv", index=False)
